Fix dot file restore errors
- Don't try to create backup for non-existant dot files
  (some files like .vimrc don't exist by default)
- Exclude .git repo from dot-files list

+++ restore.py
@@ -14,12 +14,13 @@ dot_file_paths = glob.glob(dot_files_dir + '/.*')
 for dot_file_path in dot_file_paths:
     filename = os.path.basename(dot_file_path)
     link = os.path.join(home, filename)
-
-    # If the dot file in the home directory is not already a link
-    # then it is most likely the original file.
-    if not os.path.islink(link):
-        backup_file = os.path.join(home, filename + "_backup")
-        print('Creating back up file', backup_file)
-        shutil.move(link, backup_file)
-        print('Creating symlink', link, '->', dot_file_path)
-        os.symlink(dot_file_path, link)
+    if filename != ".git":
+        # If the dot file in the home directory is not already a link
+        # then it is most likely the original file.
+        if not os.path.islink(link):
+            if os.path.isfile(link):
+                backup_file = os.path.join(home, filename + "_backup")
+                print('Creating back up file', backup_file)
+                shutil.move(link, backup_file)
+            print('Creating symlink', link, '->', dot_file_path)
+            os.symlink(dot_file_path, link)
